make FlacStream take the reader by value
This is strictly more general, because you can take a `&mut io::Reader` by
value. I am hoping this will increase performance via devirtualisation, but so
far it appears that performance has only decreased. The internal frame readers
still need to be converted though, but it raises a question:

If `reader: R` where `R: io::Reader`, then what is the type of `&mut reader`?
Is is `&mut R`, or is it `&mut io::Reader`? I am afraid it might be the latter,
at least when it is passed into `fn f<R: io::Reader>(r: R)` due to coercion,
because there is no impl `io::Reader` for `&mut R` where `R: io::Reader`. If
this is the case, then perhaps it can be fixed.