Try to see actual times through the noise
I want to put proper error bars on my values, but I don't know how to do
that in this case.

I assume that for a particular block, there is a true decode time T,
and I can observe T + noise, where the noise is drawn from some
distribution. It looks steeper than exponential. Perhaps a gamma
distribution with k < 1.0. So then the question is: given a number of
observations of T + noise, give the maximum likelihood estimate for T
and the parameters of the noise distribution. I should work this out
with pen and paper; now that I am writing it down it seems doable after
all.