Report mean time and confidence intervals
The confidence interval estimate is a bit ad-hoc, but at least
empirically it appears to work well. I have a few different runs of the
benchmark (without changing the program) to judge variance, and all of
my measurements are within the confidence bounds reported by the others.
Also, one measurement was taken while my machine was busy, and the
estimate for the time per sample is very close there, but the confidence
interval is much wider, so it appears to work.